+++ crates/ruma-common/tests/it/identifiers/ui/01-valid-id-macros.rs
@@ -7,7 +7,7 @@ fn main() {
     _ = ruma_common::room_id!("!1234567890:matrix.org");
     _ = ruma_common::room_version_id!("1");
     _ = ruma_common::room_version_id!("1-custom");
-    _ = ruma_common::server_signing_key_id!("ed25519:Abc_1");
+    _ = ruma_common::server_signing_key_version!("Abc_1");
     _ = ruma_common::server_name!("myserver.fish");
     _ = ruma_common::user_id!("@user:ruma.io");